mobile/android/tests/browser/chrome/test_session_parentid.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1301160
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1301160</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript">

  /** Test for Bug 1301160 **/

  "use strict";

  ChromeUtils.import("resource://gre/modules/XPCOMUtils.jsm");
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");
  ChromeUtils.import("resource://gre/modules/Messaging.jsm");

  // The chrome window and related services
  let chromeWin = Services.wm.getMostRecentWindow("navigator:browser");
  let BrowserApp = chromeWin.BrowserApp;
  let ss = Cc["@mozilla.org/browser/sessionstore;1"].getService(Ci.nsISessionStore);

  // Track the tabs where the tests are happening
  let tabParent;
  let tabChild1;
  let tabChild2;

  function cleanupTabs() {
    if (tabParent) {
      BrowserApp.closeTab(tabParent);
      tabParent = null;
    }

    if (tabChild1) {
      BrowserApp.closeTab(tabChild1);
      tabChild1 = null;
    }

    if (tabChild2) {
      BrowserApp.closeTab(tabChild2);
      tabChild2 = null;
    }
  }

  const url = "data:text/html;charset=utf-8,It%20was%20a%20dark%20and%20stormy%20night.";

  add_task(async function test_sessionStoreParentId() {
    SimpleTest.registerCleanupFunction(function() {
      cleanupTabs();
    });

    // Initialize parent tab
    tabParent = BrowserApp.addTab(url, { selected: true });
    await promiseTabEvent(tabParent.browser, "DOMContentLoaded");
    is(BrowserApp.selectedTab, tabParent, "tabParent is selected");

    // test case #1
    // Open tabs without passing a parent tab ID
    tabChild1 = BrowserApp.addTab(url, { selected: false });
    await promiseTabEvent(tabChild1.browser, "DOMContentLoaded");
    is(BrowserApp.selectedTab, tabParent, "tabParent is selected");

    tabChild2 = BrowserApp.addTab(url, { selected: true });
    await promiseTabEvent(BrowserApp.deck, "TabSelect");
    is(BrowserApp.selectedTab, tabChild2, "2nd child tab is selected");

    // After closing that tab, its neighbour should be selected
    BrowserApp.closeTab(tabChild2);
    tabChild2 = null;
    await promiseTabEvent(BrowserApp.deck, "TabClose");
    is(BrowserApp.selectedTab, tabChild1, "1st child tab is selected");

    // test case #2
    // Let's open a new tab, this time with a parent id and let's check if after closing it,
    // the selected tab will be the neighbour since the parent was not selected
    tabChild2 = BrowserApp.addTab(url, { selected: false, parentId: tabParent.id });
    await promiseTabEvent(tabChild2.browser, "DOMContentLoaded");
    is(BrowserApp.selectedTab, tabChild1, "1st child tab is still selected");

    BrowserApp.selectTab(tabChild2);
    await promiseTabEvent(BrowserApp.deck, "TabSelect");
    is(BrowserApp.selectedTab, tabChild2, "2nd child tab is selected");

    BrowserApp.closeTab(tabChild2);
    tabChild2 = null;
    await promiseTabEvent(BrowserApp.deck, "TabClose");
    is(BrowserApp.selectedTab, tabChild1, "1st child tab is selected");

    // test case #3
    // This time we open a new tab with a parent id but this time the parent should be selected
    // after closing since the parent was the previously selected tab
    BrowserApp.selectTab(tabParent);
    await promiseTabEvent(BrowserApp.deck, "TabSelect");
    is(BrowserApp.selectedTab, tabParent, "tabParent is selected");

    tabChild2 = BrowserApp.addTab(url, { selected: true, parentId: tabParent.id });
    await promiseTabEvent(tabChild2.browser, "DOMContentLoaded");
    is(BrowserApp.selectedTab, tabChild2, "2d child tab is selected");

    BrowserApp.closeTab(tabChild2);
    tabChild2 = null;
    await promiseTabEvent(BrowserApp.deck, "TabClose");
    is(BrowserApp.selectedTab, tabParent, "tabParent is selected");

    // test case #4
    // Now check that this works even if the child tab is closed and subsequently restored
    tabChild2 = BrowserApp.addTab(url, { selected: false, parentId: tabParent.id });
    await promiseTabEvent(tabChild2.browser, "SSTabDataUpdated");
    BrowserApp.closeTab(tabChild2);
    await promiseTabEvent(tabChild2.browser, "SSTabCloseProcessed");

    // Restore the tab
    let closedTabData = ss.getClosedTabs(chromeWin)[0];
    let browser = ss.undoCloseTab(chromeWin, closedTabData);
    await promiseTabEvent(BrowserApp.deck, "TabSelect");
    tabChild2 = BrowserApp.getTabForBrowser(browser);
    is(BrowserApp.selectedTab, tabChild2, "restored 2nd child tab is selected");

    // After closing that tab, its parent should be selected
    BrowserApp.closeTab(tabChild2);
    tabChild2 = null;
    await promiseTabEvent(BrowserApp.deck, "TabSelect");
    is(BrowserApp.selectedTab, tabParent, "tabParent is selected after restoring");

    cleanupTabs();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1301160">Mozilla Bug 1301160</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
